requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
schedule==1.2.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
